Refactor Illuminate\Foundation\Application\TestCase to AssertionTrait and ApplicationTrait
This allow easier testing support for developer wanting to use testing facility other than PHPUnit such as Behat or etc.
